# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Beauty Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{id}/notes:
    get:
      operationId: listPatientNotes
      tags: [patient-notes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientNote"
    post:
      operationId: createPatientNote
      tags: [patient-notes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientNote"
  /patient-notes/{id}:
    delete:
      operationId: deletePatientNote
      tags: [patient-notes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Services ──────────────────────────────────────────────────────────────
...
  /discounts:
    get:
      operationId: listDiscounts
      tags: [discounts]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discount"
    post:
      operationId: createDiscount
      tags: [discounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discount"
  /discounts/{id}:
    put:
      operationId: updateDiscount
      tags: [discounts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discount"
        "404":
          description: Not found
    delete:
      operationId: deleteDiscount
      tags: [discounts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Inventory ──────────────────────────────────────────────────────────────
...
  /commissions:
    get:
      operationId: listCommissions
      tags: [commissions]
      parameters:
        - name: isPaid
          in: query
          schema: { type: boolean }
        - name: recipientType
          in: query
          schema: { type: string }
        - name: recipientId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commission"
    post:
      operationId: createCommission
      tags: [commissions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"
  /commissions/{id}:
    put:
      operationId: updateCommission
      tags: [commissions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"
        "404":
          description: Not found
    delete:
      operationId: deleteCommission
      tags: [commissions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Commission Recipients ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Discount:
      type: object
      required: [id, name, code, type, value, usageCount, isActive, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        type: { type: string }
        value: { type: integer }
        minAmount: { type: ["integer", "null"] }
        usageLimit: { type: ["integer", "null"] }
        usageCount: { type: integer }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        isActive: { type: boolean }
        description: { type: ["string", "null"] }
        createdAt: { type: integer }
    DiscountInput:
      type: object
      required: [name, code, type, value]
      properties:
        name: { type: string }
        code: { type: string }
        type: { type: string }
        value: { type: integer }
        minAmount: { type: integer }
        usageLimit: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
        isActive: { type: boolean }
        description: { type: string }
    DiscountUpdate:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
        type: { type: string }
        value: { type: integer }
        minAmount: { type: ["integer", "null"] }
        usageLimit: { type: ["integer", "null"] }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        isActive: { type: boolean }
        description: { type: ["string", "null"] }

    # ── Inventory ─────────────────────────────────────────────────────────────
...
    Commission:
      type: object
      required: [id, recipientType, recipientId, amount, isPaid, createdAt]
      properties:
        id: { type: integer }
        recipientType: { type: string }
        recipientId: { type: integer }
        appointmentId: { type: ["integer", "null"] }
        paymentId: { type: ["integer", "null"] }
        description: { type: ["string", "null"] }
        amount: { type: integer }
        rate: { type: ["integer", "null"] }
        status: { type: string }
        isPaid: { type: boolean }
        paidAt: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: integer }
        recipientName: { type: ["string", "null"] }
    CommissionInput:
      type: object
      required: [recipientType, recipientId, amount]
      properties:
        recipientType: { type: string }
        recipientId: { type: integer }
        appointmentId: { type: integer }
        paymentId: { type: integer }
        description: { type: string }
        amount: { type: integer }
        rate: { type: integer }
        status: { type: string }
        notes: { type: string }
    CommissionUpdate:
      type: object
      properties:
        recipientType: { type: string }
        recipientId: { type: integer }
        appointmentId: { type: ["integer", "null"] }
        description: { type: ["string", "null"] }
        amount: { type: integer }
        rate: { type: ["integer", "null"] }
        status: { type: string }
        isPaid: { type: boolean }
        paidAt: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }

    # ── Patient Note ──────────────────────────────────────────────────────────
    PatientNote:
      type: object
      required: [id, patientId, content, kind, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        content: { type: string }
        kind: { type: string }
        createdAt: { type: integer }
    PatientNoteInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
        kind: { type: string }

    # ── Reminder ──────────────────────────────────────────────────────────────
...